chore: move files
move files from the state folder to the root src/ folder
rename store/index.js to store.js
