+++ webapp/src/constants/ActionTypes.js
@@ -1,0 +1,2 @@
+/* eslint-disable import/prefer-default-export */
+export const FOOTER_BUTTON = 'FOOTER_BUTTON'
