+++ webapp/src/reducers/footer.js
