--- webapp/src/state/constants/ActionTypes.js
@@ -1,2 +1,0 @@
-/* eslint-disable */
-export const FOOTER_BUTTON = 'FOOTER_BUTTON'
